+++ src/dll/Addresses.hpp
@@ -14,7 +14,7 @@ public:
 
     ~Addresses() = default;
 
-    std::uintptr_t Resolve(const RED4ext::UniversalRelocSegment aSegment, const std::uint64_t aHash) const;
+    std::uintptr_t Resolve(const RED4ext::UniversalRelocSegment aSegment, const std::uint32_t aHash) const;
 
 private:
     Addresses(const Paths& aPaths);
@@ -22,5 +22,5 @@ private:
     void LoadAddresses(const std::filesystem::path& aPath);
 
     std::uintptr_t m_codeOffset;
-    std::unordered_map<std::uint64_t, std::uintptr_t> m_addresses;
+    std::unordered_map<std::uint32_t, std::uintptr_t> m_addresses;
 };
